Auto-migrate to MUI 5
Runs the `preset-safe` codemod from the migration guide at
https://mui.com/material-ui/migration/migration-v4/.

There are some CSS inheritance regressions, namely with `primary` color
styles not being overwritten to default or `secondary` states (a common
example would be in the `<BackButton>`.
There are also some spacing/sizing inconsistencies that need investigation.
A rewrite is needed to migrate from JSS to a modern emotion engine syntax.
Once the dependency on `@mui/styles` is removed (through this rewrite),
it will be possible to update to the latest React 18.

Further, some breaking changes mentioned in the migration guide need
manual intervention, which has not yet been completed.